LPS-80017 Extract UpgradeExecutor class from ReleaseManagerOSGiCommands, no logic change.
