+++ modules/apps/portal/portal-upgrade-impl/src/main/java/com/liferay/portal/upgrade/internal/release/osgi/commands/ReleaseManagerOSGiCommands.java
@@ -19,32 +19,21 @@ import com.liferay.osgi.service.tracker.collections.map.PropertyServiceReference
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapListener;
-import com.liferay.petra.reflect.ReflectionUtil;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-import com.liferay.portal.kernel.cache.CacheRegistryUtil;
 import com.liferay.portal.kernel.configuration.Configuration;
 import com.liferay.portal.kernel.configuration.ConfigurationFactoryUtil;
 import com.liferay.portal.kernel.dao.db.DB;
-import com.liferay.portal.kernel.dao.db.DBContext;
 import com.liferay.portal.kernel.dao.db.DBManagerUtil;
-import com.liferay.portal.kernel.dao.db.DBProcessContext;
 import com.liferay.portal.kernel.model.Release;
-import com.liferay.portal.kernel.model.ReleaseConstants;
 import com.liferay.portal.kernel.service.ReleaseLocalService;
 import com.liferay.portal.kernel.upgrade.UpgradeStep;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.StringBundler;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.output.stream.container.OutputStreamContainer;
-import com.liferay.portal.output.stream.container.OutputStreamContainerFactory;
-import com.liferay.portal.output.stream.container.OutputStreamContainerFactoryTracker;
 import com.liferay.portal.upgrade.internal.configuration.ReleaseManagerConfiguration;
+import com.liferay.portal.upgrade.internal.executor.UpgradeExecutor;
 import com.liferay.portal.upgrade.internal.graph.ReleaseGraphManager;
 import com.liferay.portal.upgrade.internal.registry.UpgradeInfo;
-import com.liferay.portal.upgrade.internal.release.ReleasePublisher;
-
-import java.io.IOException;
-import java.io.OutputStream;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -151,7 +140,7 @@ public class ReleaseManagerOSGiCommands {
 		ReleaseGraphManager releaseGraphManager = new ReleaseGraphManager(
 			_serviceTrackerMap.getService(bundleSymbolicName));
 
-		executeUpgradeInfos(
+		_upgradeExecutor.executeUpgradeInfos(
 			bundleSymbolicName,
 			releaseGraphManager.getUpgradeInfos(
 				schemaVersionString, toVersionString));
@@ -208,15 +197,6 @@ public class ReleaseManagerOSGiCommands {
 		for (UpgradeInfo upgradeProcess : upgradeProcesses) {
 			System.out.println("\t" + upgradeProcess);
 		}
-	}
-
-	@Reference(unbind = "-")
-	public void setOutputStreamTracker(
-		OutputStreamContainerFactoryTracker
-			outputStreamContainerFactoryTracker) {
-
-		_outputStreamContainerFactoryTracker =
-			outputStreamContainerFactoryTracker;
 	}
 
 	@Activate
@@ -278,7 +258,8 @@ public class ReleaseManagerOSGiCommands {
 			return;
 		}
 
-		executeUpgradeInfos(bundleSymbolicName, upgradeInfosList.get(0));
+		_upgradeExecutor.executeUpgradeInfos(
+			bundleSymbolicName, upgradeInfosList.get(0));
 	}
 
 	protected void executeAll(
@@ -314,44 +295,6 @@ public class ReleaseManagerOSGiCommands {
 		}
 
 		executeAll(upgradeThrewExceptionBundleSymbolicNames);
-	}
-
-	protected void executeUpgradeInfos(
-		final String bundleSymbolicName, final List<UpgradeInfo> upgradeInfos) {
-
-		OutputStreamContainerFactory outputStreamContainerFactory =
-			_outputStreamContainerFactoryTracker.
-				getOutputStreamContainerFactory();
-
-		OutputStreamContainer outputStreamContainer =
-			outputStreamContainerFactory.create(
-				"upgrade-" + bundleSymbolicName);
-
-		OutputStream outputStream = outputStreamContainer.getOutputStream();
-
-		Release release = _releaseLocalService.fetchRelease(bundleSymbolicName);
-
-		if (release != null) {
-			_releasePublisher.publishInProgress(release);
-		}
-
-		_outputStreamContainerFactoryTracker.runWithSwappedLog(
-			new UpgradeInfosRunnable(
-				bundleSymbolicName, upgradeInfos, outputStream),
-			outputStreamContainer.getDescription(), outputStream);
-
-		try {
-			outputStream.close();
-		}
-		catch (IOException ioe) {
-			throw new RuntimeException(ioe);
-		}
-
-		release = _releaseLocalService.fetchRelease(bundleSymbolicName);
-
-		if (release != null) {
-			_releasePublisher.publish(release);
-		}
 	}
 
 	protected String getSchemaVersionString(String bundleSymbolicName) {
@@ -406,19 +349,14 @@ public class ReleaseManagerOSGiCommands {
 		_releaseLocalService = releaseLocalService;
 	}
 
-	@Reference(unbind = "-")
-	protected void setReleasePublisher(ReleasePublisher releasePublisher) {
-		_releasePublisher = releasePublisher;
-	}
-
 	private static Logger _logger;
 
-	private OutputStreamContainerFactoryTracker
-		_outputStreamContainerFactoryTracker;
 	private ReleaseLocalService _releaseLocalService;
 	private ReleaseManagerConfiguration _releaseManagerConfiguration;
-	private ReleasePublisher _releasePublisher;
 	private ServiceTrackerMap<String, List<UpgradeInfo>> _serviceTrackerMap;
+
+	@Reference
+	private UpgradeExecutor _upgradeExecutor;
 
 	private static class UpgradeServiceTrackerCustomizer
 		implements ServiceTrackerCustomizer<UpgradeStep, UpgradeInfo> {
@@ -512,93 +450,6 @@ public class ReleaseManagerOSGiCommands {
 			String key, UpgradeInfo upgradeInfo,
 			List<UpgradeInfo> upgradeInfos) {
 		}
-
-	}
-
-	private class UpgradeInfosRunnable implements Runnable {
-
-		public UpgradeInfosRunnable(
-			String bundleSymbolicName, List<UpgradeInfo> upgradeInfos,
-			OutputStream outputStream) {
-
-			_bundleSymbolicName = bundleSymbolicName;
-			_upgradeInfos = upgradeInfos;
-			_outputStream = outputStream;
-		}
-
-		@Override
-		public void run() {
-			int buildNumber = 0;
-			int state = ReleaseConstants.STATE_GOOD;
-
-			try {
-				_updateReleaseState(_STATE_IN_PROGRESS);
-
-				for (UpgradeInfo upgradeInfo : _upgradeInfos) {
-					UpgradeStep upgradeStep = upgradeInfo.getUpgradeStep();
-
-					upgradeStep.upgrade(
-						new DBProcessContext() {
-
-							@Override
-							public DBContext getDBContext() {
-								return new DBContext();
-							}
-
-							@Override
-							public OutputStream getOutputStream() {
-								return _outputStream;
-							}
-
-						});
-
-					_releaseLocalService.updateRelease(
-						_bundleSymbolicName,
-						upgradeInfo.getToSchemaVersionString(),
-						upgradeInfo.getFromSchemaVersionString());
-
-					buildNumber = upgradeInfo.getBuildNumber();
-				}
-			}
-			catch (Exception e) {
-				state = ReleaseConstants.STATE_UPGRADE_FAILURE;
-
-				ReflectionUtil.throwException(e);
-			}
-			finally {
-				Release release = _releaseLocalService.fetchRelease(
-					_bundleSymbolicName);
-
-				if (release != null) {
-					if (buildNumber > 0) {
-						release.setBuildNumber(buildNumber);
-					}
-
-					release.setState(state);
-
-					_releaseLocalService.updateRelease(release);
-				}
-			}
-
-			CacheRegistryUtil.clear();
-		}
-
-		private void _updateReleaseState(int state) {
-			Release release = _releaseLocalService.fetchRelease(
-				_bundleSymbolicName);
-
-			if (release != null) {
-				release.setState(state);
-
-				_releaseLocalService.updateRelease(release);
-			}
-		}
-
-		private static final int _STATE_IN_PROGRESS = -1;
-
-		private final String _bundleSymbolicName;
-		private final OutputStream _outputStream;
-		private final List<UpgradeInfo> _upgradeInfos;
 
 	}
